Atualização do arquivo base do banco de dados
Mudança no telefone, mudado para uma relação 1 - N... Criados os mecanismos para o melhor gerenciamento dos numero de telefones de cada contato

+++ project_jsp/src/DAO/DAOUser.java
@@ -20,19 +20,20 @@ public class DAOUser {
 
 	public void save(UserBean user) {
 		try {
-			String sql = "insert into usuario values(null,?,?,?,?,?,?,?,?,?,?,?)";
+			String sql = "insert into usuario values(null,?,?,?,?,?,?,?,?,?,?,?,?)";
 			PreparedStatement statement = conn.prepareStatement(sql);
 			statement.setString(1, user.getLogin());
 			statement.setString(2, user.getPassword());
 			statement.setString(3, user.getName());
-			statement.setString(4, user.getPhone());
-			statement.setString(5, user.getzCode());
-			statement.setString(6, user.getStreet());
-			statement.setString(7, user.getNumber());
-			statement.setString(8, user.getDistrict());
-			statement.setString(9, user.getCity());
-			statement.setString(10, user.getState());
-			statement.setString(11, user.getIbge());
+			statement.setString(4, user.getzCode());
+			statement.setString(5, user.getStreet());
+			statement.setString(6, user.getNumber());			
+			statement.setString(7, user.getDistrict());
+			statement.setString(8, user.getCity());
+			statement.setString(9, user.getState());
+			statement.setString(10, user.getIbge());
+			statement.setString(11, user.getPhotoBase64());
+			statement.setString(12, user.getContentType());
 			statement.execute();
 			conn.commit();
 
@@ -49,7 +50,7 @@ public class DAOUser {
 
 	public void delete(String id) {
 		try {
-			String sql = "delete from usuario where id ='" + id + "'";
+			String sql = "delete from usuario where id ='" + id + "' and login != 'admin'";
 			PreparedStatement statement = conn.prepareStatement(sql);
 			statement.execute();
 			conn.commit();
@@ -66,22 +67,23 @@ public class DAOUser {
 	public void update(UserBean user) {
 		try {
 			String sql = "update usuario set login=?,"
-					+ " senha=?, nome=?, telefone=?, cep=?,"
-					+ " rua=?, num=?, bairro=?, cidade=?, uf=?, ibge=?"
-					+ " where id =" + user.getId();
+					+ " senha=?, nome=?, cep=?,"
+					+ " rua=?, num=?, bairro=?, cidade=?, uf=?, ibge=?, foto=?, tipo=?"
+					+ " where id =" + user.getId() +" and login != 'admin'" ;
 
 			PreparedStatement statement = conn.prepareStatement(sql);
 			statement.setString(1, user.getLogin());
 			statement.setString(2, user.getPassword());
 			statement.setString(3, user.getName());
-			statement.setString(4, user.getPhone());
-			statement.setString(5, user.getzCode());
-			statement.setString(6, user.getStreet());
-			statement.setString(7, user.getNumber());			
-			statement.setString(8, user.getDistrict());
-			statement.setString(9, user.getCity());
-			statement.setString(10, user.getState());
-			statement.setString(11, user.getIbge());
+			statement.setString(4, user.getzCode());
+			statement.setString(5, user.getStreet());
+			statement.setString(6, user.getNumber());			
+			statement.setString(7, user.getDistrict());
+			statement.setString(8, user.getCity());
+			statement.setString(9, user.getState());
+			statement.setString(10, user.getIbge());
+			statement.setString(11, user.getPhotoBase64());
+			statement.setString(12, user.getContentType());
 			statement.executeUpdate();
 			conn.commit();
 		} catch (SQLException e) {
@@ -96,7 +98,7 @@ public class DAOUser {
 
 	public List<UserBean> listAll() throws SQLException {
 		List<UserBean> lst = new ArrayList<UserBean>();
-		String sql = "select * from usuario";
+		String sql = "select * from usuario where login != 'admin'";
 
 		PreparedStatement statement = conn.prepareStatement(sql);
 		ResultSet res = statement.executeQuery();
@@ -104,25 +106,27 @@ public class DAOUser {
 		while (res.next()) {
 			UserBean user = new UserBean(
 					res.getLong("id"), res.getString("login"), res.getString("senha"),
-					res.getString("nome"), res.getString("telefone"), res.getString("cep"),
+					res.getString("nome"), res.getString("cep"),
 					res.getString("rua"), res.getString("num"),res.getString("bairro"),
-					res.getString("cidade"), res.getString("uf"),res.getString("ibge"));
+					res.getString("cidade"), res.getString("uf"),res.getString("ibge"),
+					res.getString("foto"),res.getString("tipo"));
 			lst.add(user);
 		}
 		return lst;
 	}
 
 	public UserBean findById(String id) throws SQLException {
-		String sql = "select * from usuario where id='" + id + "'";
+		String sql = "select * from usuario where id='" + id + "' and login != 'admin'";
 		PreparedStatement statement = conn.prepareStatement(sql);
 		ResultSet res = statement.executeQuery();
 
 		if (res.next()) {
 			UserBean user = new UserBean(
 					res.getLong("id"), res.getString("login"), res.getString("senha"),
-					res.getString("nome"), res.getString("telefone"), res.getString("cep"),
+					res.getString("nome"), res.getString("cep"),
 					res.getString("rua"), res.getString("num"),res.getString("bairro"),
-					res.getString("cidade"), res.getString("uf"),res.getString("ibge"));
+					res.getString("cidade"), res.getString("uf"),res.getString("ibge"),
+					res.getString("foto"),res.getString("tipo"));
 			return user;
 		}
 		return null;
